Alteração do tipo de coluna 'group' da tabela 'equivalent_foods'

+++ db/migrate/20150304031442_create_equivalent_foods.rb
@@ -3,9 +3,9 @@ class CreateEquivalentFoods < ActiveRecord::Migration
     create_table :equivalent_foods do |t|
       t.string :food
       t.integer :group
-      t.integer :quantity_measure
+      t.float :quantity_measure
       t.string :measure
-      t.integer :weight
+      t.float :weight
 
       t.timestamps null: false
     end
